Add files via upload
REAP (Renormalization group Evolution of Angles and Phases) is a Mathematica based package for solving the renormalization group equations (RGE) of the quantities relevant for neutrino masses (https://reapmpt.hepforge.org). It includes several models, one of which is the MSSM0N model. This model runs the RGs from the unification scale down to the electroweak scale using the MSSM. In the electroweak scale the neutrino mixing parameters can be predicted based on the free parameters set at the unification scale, for example in the framework of the HSMU model (Phys. Rev. D 69, 053007 (2004), hep-ph/0301234).

In this project, the MSSM0N model has been automatized via neural networks using MLP via multivariate regression, for the case of no Threshold Corrections. This can be extended to all models included in the REAP package and regarding the amount of time Mathematica takes for running, it can save a huge time for operators of the REAP code.

A short data file is uploaded as a demo. The columns include 10 initial parameters at unification scale (Features) and 12 observables at electroweak scale(Lables)